Update index.html
I like this version of the "home page/landing page" better.
It's more responsive thsn the previous edit. And since the stylesheet messed up the responsiveness of this file, we moved the style to be internal for better usage 

TO DO NEXT: 
Add button to link all html files together. 
Add "<script src="script.js"></script>" to the end instead of having a internal JS. either is fine too ig.